Remove base class
